Revert "Styling"
