+++ stock-tracker/src/App.js
@@ -4,27 +4,32 @@ import KeyStats from "./components/keyStats";
 import Chart from "./components/chart";
 import LatestNews from "./components/LatestNews";
 import Overview from "./components/Overview";
+import TopPeers from "./components/TopPeers";
 import Footer from "./components/Footer";
 import "./assets/styles/App.css";
 import { useSelector } from "react-redux";
 
 function App() {
   const symbol = useSelector(state => state.symbol);
-  return (
-    <>
-      <div className="wrapper">
+  if (symbol === "") {
+    return (
+      <div className="wrapper_default">
         <Header />
-        {symbol !== "" && (
-          <>
-            <Chart />
-            <KeyStats />
-            <LatestNews />
-            <Overview />
-          </>
-        )}
+        <span className="content"></span>
+        <Footer />
       </div>
-      {symbol !== "" && <Footer />}
-    </>
+    );
+  }
+  return (
+    <div className="wrapper">
+      <Header />
+      <Chart />
+      <KeyStats />
+      <LatestNews />
+      <Overview />
+      <TopPeers />
+      <Footer />
+    </div>
   );
 }
 export default App;
