+++ stock-tracker/src/assets/styles/Header.css
@@ -1,30 +1,28 @@
 .header {
   display: grid;
-  grid-template-columns: 2fr 1.5fr;
+  grid-template-columns: 2fr 1fr;
   grid-template-areas:
-    "header__logo header__btns"
-    "search_price_container search_price_container"
-    "currency_display market_status_display";
+    "logo headerBtns"
+    "searchDisplay priceDisplay"
+    "currencyDisplay marketStatusDisplay";
 }
 .header__logo {
-  grid-area: header__logo;
+  grid-area: logo;
 }
+
 .header__btns {
-  grid-area: header__btns;
+  grid-area: headerBtns;
+}
+.search_display {
+  grid-area: searchDisplay;
+}
+
+.price_display {
+  grid-area: priceDisplay;
 }
 .currency_display {
-  grid-area: currency_display;
+  grid-area: currencyDisplay;
 }
 .market_status_display {
-  grid-area: market_status_display;
-}
-@media (max-width: 425px) {
-  .header {
-    grid-template-columns: 1fr 2fr;
-  }
-}
-.header__search_price_container {
-  display: grid;
-  grid-area: search_price_container;
-  grid-template-columns: 2fr 1fr;
+  grid-area: marketStatusDisplay;
 }
