+++ stock-tracker/src/components/LatestNews.js
@@ -7,13 +7,13 @@ const LatestNews = () => {
   const latestNews = useSelector(state => state.latestNews);
   const loading = useSelector(state => state.loading);
   const newsDisplay = latestNews.map((news, index) => (
-    <div className="news" key={index}>
-      <p className={index === 0 ? "first_news news_headline" : "news_headline"}>
+    <div key={index}>
+      <p className={index === 0 ? "firstNews newsHeadline" : "newsHeadline"}>
         {" "}
         <a href={news.url}>{news.headline}</a>
       </p>
-      <p className="news_time">{moment(news.datetime).fromNow()}</p>
-      <p className="news_source">- {news.source}</p>
+      <p className="newsTime">{moment(news.datetime).fromNow()}</p>
+      <p className="newsSource">- {news.source}</p>
     </div>
   ));
   const errorValue = useSelector(state => state.error.latestNews);
